code restructure: rename components
