+++ src/components/fetch/FetchRouteList.tsx
@@ -3,7 +3,7 @@ import { Card } from "@fluentui/react-components/unstable";
 import { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 
-import { RoutesXml } from "../../data/etaXml";
+import { RoutesXml } from "../../models/etaXml";
 import RawDisplay from "../rawDisplay/RawDisplay";
 import { FetchXMLWithCancelToken } from "./fetchUtils";
